tcp client refactor
